Added redux

+++ backend/controllers/response.js
@@ -5,6 +5,7 @@ const Response = require("../models/Response");
 const User = require("../models/User");
 const mongoose = require("mongoose");
 const mailer = require('../config/mail');
+const Form = require("../models/Form");
 
 const addResponse = asyncHandler(async (req, res) => {
   const { formId } = req.params;
@@ -16,7 +17,10 @@ const addResponse = asyncHandler(async (req, res) => {
   const response = await Response.create({ form: formId, user: userId });
   if (response) {
     // get list of questions
-    const questions = await Question.find({ form: formId }).lean();
+    const questions = await Question.find({ form: formId }).populate('form').lean();
+    const form = questions && questions[0].form;
+    const formOwnerId = form && form.user;
+    console.log(formOwnerId,form)
     //   req.body["questionId"] = answerText
     const data = questions.map((question) => {
       return {
@@ -27,8 +31,9 @@ const addResponse = asyncHandler(async (req, res) => {
     });
     const answers = await Answer.insertMany(data);
     if (answers) {
-      res.status(201).json([ ...answers ]);
       const user = await User.findById(req.user.id).select('name email').lean();
+
+    // Send mail to user who filled the form
       const list = questions.map(item=>`<li>
         <p>${item.text}</p>
         <p>${req.body[item._id] || "Not filled"}</p>
@@ -42,6 +47,20 @@ const addResponse = asyncHandler(async (req, res) => {
         html: responseHTML,
       };
      await mailer.send(message)
+
+    // Send mail to user who owns the form
+      const formOwner = await User.findById(formOwnerId).select('name email').lean();
+      const OwnerResponse = `<h1>${user.name} has submitted a resposonse </h1> <ul>${list}</ul>`
+      console.log(formOwner)
+      const ownerMsg = {
+        to: formOwner.email,
+        from: "sxgean@block521.com",
+        subject: "Your form got a response!",
+        text: `${user.name} submitted a response!`,
+        html: OwnerResponse,
+      };
+      await mailer.send(ownerMsg);
+        res.status(201).json([...answers]);
     } else {
       res.status(400);
       throw new Error("Something went wrong");
@@ -104,15 +123,22 @@ const deleteResponse = asyncHandler(async (req, res) => {
       throw new Error("Invalid Form Id");
     }
     const answers = await Answer.find({ response: responseId }).select('_id').lean()
-    const deleteIds = answers.map(answer=>answer._id)
-    await Answer.deleteMany({_id:{$in:deleteIds}})
-    const deletedResponse = await Response.findByIdAndDelete(responseId).lean()
-    if(deleteResponse){
-        res.status(200).json({...deletedResponse})
-    }
-    else{
-        res.status(400);
-        throw new Error("Something went wrong!");
+    const response = await Response.findById(responseId).lean()
+    const userId  = response && response.user;
+    if(userId == req.user.id){
+      const deleteIds = answers.map(answer=>answer._id)
+      await Answer.deleteMany({_id:{$in:deleteIds}})
+      const deletedResponse = await Response.findByIdAndDelete(responseId).lean()
+      if(deleteResponse){
+          res.status(200).json({...deletedResponse})
+      }
+      else{
+          res.status(400);
+          throw new Error("Something went wrong!");
+      }
+    }else{
+      res.status(403);
+      throw new Error("Only user who filled the form can delete the response.")
     }
 });
